pyproject.toml
[project]
name = "deepseek_nvidia"
version = "0.1.0"
description = "Send prompts to the NVIDIA DeepSeek AI"
authors = [{ name = "Ian Sweeney" }]
requires-python = ">=3.12"
dependencies = [
  # Add runtime dependencies here
  "numpy",
  "openai",
  "pyyaml",
  "torch",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true  # if needed for local paths or GitHub links

[tool.hatch.envs.default]
python = "3.12"
dependencies = [
  # Add dev dependencies here
  "pytest",
]

[tool.hatch.envs.default.scripts]
test = "pytest"
format = "ruff format ."
lint = "ruff check ."
type_check = "mypy ."
update = """
  pip install -U pip setuptools wheel &&
  pip list --outdated --format=freeze | cut -d = -f 1 | xargs -n1 pip install -U
"""
